requirements.txt
joblib
PyJWT
Flask
numpy
SQLAlchemy